+++ .buildkite/images/docker/push.sh
@@ -1,15 +1,11 @@
 #! /bin/bash
 ROOT=$(git rev-parse --show-toplevel)
-pushd $ROOT/.buildkite/images/docker/
+cd $ROOT/.buildkite/images/docker/
 set -eux
 
-# Version of the buildkite image. Update this when you make significant changes to the image.
-IMAGE_VERSION="v6"
-
-
-if [ "$#" -ne 2 ]; then
-    echo "Error: Must specify a Python version and image type.\n" 1>&2
-    echo "Usage: ./push.sh 3.7.4 integration" 1>&2
+if [ "$#" -ne 3 ]; then
+    echo "Error: Must specify a Python version and image type along with version.\n" 1>&2
+    echo "Usage: ./build.sh 3.7.4 integration v6" 1>&2
     exit 1
 fi
 
@@ -21,6 +17,9 @@ PYTHON_MAJOR_VERSION="${PYTHON_VERSION:0:1}"
 PYTHON_MAJMIN=`echo "${PYTHON_VERSION:0:3}" | sed 's/\.//'`
 
 IMAGE_TYPE=$2
+
+# Version of the buildkite integration image
+IMAGE_VERSION=$3
 
 TAG=`date '+%Y-%m-%d'`