+++ python_modules/dagster-celery/dagster_celery_tests/test_version.py
@@ -1,0 +1,4 @@
+def test_version():
+    from dagster_celery.version import __version__
+
+    assert __version__