+++ python_modules/libraries/dagster-postgres/dagster_postgres/run_storage/alembic/versions/da7cd32b690d_.py
@@ -1,0 +1,20 @@
+"""Base revision -- noop
+
+Revision ID: da7cd32b690d
+Revises: 
+Create Date: 2019-11-22 11:02:06.180581
+
+"""
+# revision identifiers, used by Alembic.
+revision = 'da7cd32b690d'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    pass
+
+
+def downgrade():
+    pass