+++ CHANGES.md
@@ -1,24 +1,32 @@
-
 # Changelog
 
 ## 0.7.0
 
 **Breaking**
-- `Set` and `Tuple` can no longer be used within the config system.
+
+- `dagster.Set` and `dagster.Tuple` can no longer be used within the config system.
+- The implementation of SQL-based event log storages has been consolidated,
+  which has entailed a schema change. If you have event logs stored in a
+  Postgres- or SQLite-backed event log storage, and you would like to maintain
+  access to these logs, you should run `dagster instance migrate`. To check
+  what event log storages you are using, run `dagster instance info`.
 
 ## 0.6.6
 
 **Breaking**
+
 - The `selector` argument to `PipelineDefinition` has been removed. This API made it possible to
   construct a `PipelineDefinition` in an invalid state. Use `PipelineDefinition.build_sub_pipeline`
   instead.
 
 **New**
+
 - Added the `dagster_prometheus` library, which exposes a basic Prometheus resource.
 - Dagster Airflow DAGs may now use GCS instead of S3 for storage.
 - Expanded interface for schedule management in Dagit.
-  
+
 **Dagit**
+
 - Performance improvements when loading, displaying, and editing config for large pipelines.
 - Smooth scrolling zoom in the explore tab replaces the previous two-step zoom.
 - No longer depends on internet fonts to run, allowing fully offline dev.
@@ -31,6 +39,7 @@
 - Better handling of pipelines that aren't present in the currently loaded repo.
 
 **Bugfix**
+
 - Dagster Airflow DAGs previously could crash while handling Python errors in DAG logic.
 - Step failures when running Dagster Airflow DAGs were previously not being surfaced as task
   failures in Airflow.
@@ -38,16 +47,17 @@
   solid subselection.
 - `frozenlist` and `frozendict` now pass Dagster's parameter type checks for `list` and `dict`.
 - The GraphQL playground in Dagit is now working again.
-  
+
 **Nits**
+
 - Dagit now prints its pid when it loads.
 - Third-party dependencies have been relaxed to reduce the risk of version conflicts.
 - Improvements to docs and example code.
 
-
 ## 0.6.5
 
 **Breaking**
+
 - The interface for type checks has changed. Previously the `type_check_fn` on a custom type was
   required to return None (=passed) or else raise `Failure` (=failed). Now, a `type_check_fn` may
   return `True`/`False` to indicate success/failure in the ordinary case, or else return a
@@ -64,6 +74,7 @@
   `configField`.
 
 **Bugfix**
+
 - `PresetDefinition.from_files` is now guaranteed to give identical results across all Python
   minor versions.
 - Nested composite solids with no config, but with config mapping functions, now behave as expected.
@@ -72,9 +83,11 @@
 - Improvements to Dagit interface.
 
 **New**
+
 - dagster_pyspark now supports remote execution on EMR with the `@pyspark_solid` decorator.
 
 **Nits**
+
 - Documentation has been improved.
 - The top level config field `features` in the `dagster.yaml` will no longer have any effect.
 - Third-party dependencies have been relaxed to reduce the risk of version conflicts.